refactor: move command logic to binary executor with injectable IO
Splits the binary into a minimal entry point and a dedicated executor
module. All command logic migrates from free functions in `main.rs` to
methods on `Executor` that accept `&mut Io`, enabling output capture for
future tests and alternate frontends. The `VendorWorktree` impl and its
helpers move from the library's `exe` module into `lib.rs` alongside
`VendorRepository`, eliminating the library-side `exe` module. The
worktree test file is renamed to `attr_tests.rs` to match its scope.

refactor: move `VendorWorktree` impl from `exe.rs` into `lib.rs`
refactor: introduce `Executor` struct and `Io` for injectable output
refactor: replace `std::process::exit` on conflict with `ConflictExit`
refactor: rename `exe_tests.rs` to `attr_tests.rs`
